requirements.txt
streamlit
google-genai
tiktoken 
chromadb
langchain
langchain-community
langchain-google-genai
GitPython
python-dotenv